+++ src/main/kotlin/com/github/saundefined/bitrix_idea/activity/NotificationStartupActivity.kt
@@ -1,29 +1,67 @@
-package com.github.saundefined.bitrix_idea.activity
+package com.github.saundefined.bitrix_idea
 
 import com.github.saundefined.bitrix_idea.BitrixIdeaBundle.message
-import icons.license.CheckLicense
-import com.intellij.notification.Notification
-import com.intellij.notification.NotificationType
-import com.intellij.notification.Notifications
+import com.intellij.notification.*
+import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
+import icons.license.CheckLicense
+import org.jetbrains.annotations.NotNull
+import java.awt.*
+import java.io.IOException
+import java.net.URI
+import java.net.URISyntaxException
 
 class NotificationStartupActivity : StartupActivity {
+
+    @NotNull
+    private val NOTIFICATION_GROUP = "Bitrix Idea Plugin"
+
+    private val GITHUB_URL = "https://github.com/saundefined/bitrix-idea"
+
     override fun runActivity(project: Project) {
         val app = ApplicationManager.getApplication()
         app.invokeLater {
             if (CheckLicense.isLicensed == false) {
-                val notification = Notification(
-                    "Bitrix Idea Plugin",
-                    message("startup_notification_title"),
-                    message("startup_notification_body"),
-                    NotificationType.INFORMATION
-                )
+                val notification =
+                    Notification(NOTIFICATION_GROUP, message("startup.notification.title"), NotificationType.IDE_UPDATE)
+                notification.addAction(object :
+                    NotificationAction(message("startup.notification.action")) {
+                    override fun actionPerformed(e: AnActionEvent, notification: Notification) {
+                        openUrl(GITHUB_URL)
+                    }
+                })
+
+                notification.addAction(object : NotificationAction(message("startup.notification.dismiss")) {
+                    override fun actionPerformed(e: AnActionEvent, notification: Notification) {
+                        val notificationSettings = NotificationsConfiguration.getNotificationsConfiguration()
+                        notificationSettings.changeSettings(
+                            NOTIFICATION_GROUP,
+                            NotificationDisplayType.NONE,
+                            false,
+                            false
+                        )
+                        notification.expire()
+                    }
+                })
 
                 Notifications.Bus.notify(notification)
             }
         }
     }
 
+    private fun openUrl(url: String) {
+        if (Desktop.isDesktopSupported()) {
+            val desktop: Desktop = Desktop.getDesktop()
+            if (desktop.isSupported(Desktop.Action.BROWSE)) {
+                try {
+                    val uri = URI(url)
+                    desktop.browse(uri)
+                } catch (ignored: URISyntaxException) {
+                } catch (ignored: IOException) {
+                }
+            }
+        }
+    }
 }