+++ CHANGELOG.md
@@ -1,13 +1,35 @@
 # CHANGELOG
 
 ## [Unreleased]
+
+### Changed
+
+- Add "Don't show again" action to startup notification
+
+## [2023.4.0]
+
+### Changed
+
+- Updated Plugin Core
+- Updated Bitrix CMS core
+
+## [2023.3.0]
+
+### Changed
+
 - Updated Plugin Core
 - Updated Bitrix CMS core
 
 ## [2023.2.0]
+
+### Changed
+
 - Updated Plugin Core
 - Updated Bitrix CMS core
 
 ## [2023.1.0]
+
+### Changed
+
 - Updated Plugin Core
 - Updated Bitrix CMS core